atomichash: initial commit and metrics intergration
Adds a basic atomic hashmap library. This has speed advantages for
hashmaps where primary access pattern is accessing existing entries
at the cost of slower insert and without the ability to resize the
hashtable.

This is useful for shared hashmaps with interior mutability, eg
where values are atomics and the number of keys can be estimated
at compile time.

Adds a `StaticMetrics` struct to the metrics crate which uses a
builder pattern to make use of the atomic hashmap.